+++ src/Node.cpp
@@ -1,7 +1,7 @@
 #include "includes.h"
 #include "Node.h"
 
-Node::Node(std::vector<int> coord, std::vector<int> state, std::vector<int> modifiers) {
+Node::Node(std::vector<int> coord, std::vector<double> state, std::vector<double> modifiers) {
     this->coord = coord;
     this->state = state;
     this->modifiers = modifiers;
@@ -9,7 +9,7 @@ Node::Node(std::vector<int> coord, std::vector<int> state, std::vector<int> modi
 
 // * Getters:
 std::vector<int> Node::getCoord() {return this->coord;}
-std::vector<int> Node::getState() {return this->state;}
-std::vector<int> Node::getModifiers() {return this->modifiers;}
+std::vector<double> Node::getState() {return this->state;}
+std::vector<double> Node::getModifiers() {return this->modifiers;}
 
 double Node::getWeight() {return (double)this->state.at(0);}
